Improve detection of updates with no changesets
Modified createChangesets to inspect oldrev, newrev and refname
arguments so we can detect if the change is a create/update/delete and
whether we're operating on a tag, annotated tag, branch or tracking
branch.  If the change is a delete or we're not operating on a branch,
createChangesets returns with an empty Changeset slice and no error.