Improvements to the 'lh export' command
Support fetching user information in 'lh export' command.  Fix
attachment downloads in 'lh export' command by not including
Lighthouse credentials when making a request to a non-Lighthouse
server (for example an S3 URL used to fetch a ticket attachment).  Add
support for parsing X-Rate-Limit-Retry-After header in 429 response
and using that to sleep and then retry the rate-limited request.