Support truncated changesets
If a changeset is extremely long, Lighthouse appears to truncate the
list of changes which can result in it returning a 1-element change
array containing only the operation [<op>] rather than the normal
2-element change array [<op>, <path>] containing both the operation
and path.  We now support both, with a 1-element change array
resulting in a Change with an empty path.